+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - Fix validation "0", "1", "X" in yosys netlist
  - Fix value type in _port.py (int, "X" or None is valid)
  - Remove None from value type in _port.py (int, "X" is now valid)
+ - Optimize simulation with the help of Gemeni
 
 ## v0.12.0
  - Use default audio output